Rename decks package to playlist
